clean up splash activity
